requirements.txt
flask
flask-cors
tensorflow-cpu==2.15.0
pillow
rembg
onnxruntime
gdown
gunicorn